this approach is too annoying -- switching to use subprocess instead of fork...

+++ two/demos/sock2.py
@@ -7,26 +7,28 @@ class Python(object):
     The second process can horribly crash (e.g., segfault), and the
     controlling process will see that.
     """
-    def __init__(self, namespace=None, bufsize=4): # TODO: increase 4; make small just for testing
+    def __init__(self, namespace=None, bufsize=1024): 
         self._bufsize = bufsize
         self._namespace = globals() if namespace is None else namespace
         self._sp = socket.socketpair()
         self._fifo_name = tempfile.mktemp()
+        self._fifo_stderr_name = tempfile.mktemp()        
         self._stream = ''
-        if os.path.exists(self._fifo_name):
-            os.unlink(self._fifo_name)
         os.mkfifo(self._fifo_name)
+        os.mkfifo(self._fifo_stderr_name)
         p = os.fork()
         if not p:
             # child
             self._fifo = open(self._fifo_name, 'w+')
+            self._fifo_stderr = open(self._fifo_stderr_name, 'w+')            
             os.dup2(self._fifo.fileno(), sys.stdout.fileno())
+            os.dup2(self._fifo_stderr.fileno(), sys.stderr.fileno())
             self._child_runloop0()
         else:
             # parent
             self._child_pid = p
-            #self._fifo = os.open(self._fifo_name, os.O_RDONLY|os.O_NONBLOCK)
-            self._fifo = os.open(self._fifo_name, os.O_RDONLY)
+            self._fifo = os.open(self._fifo_name, os.O_RDONLY|os.O_NONBLOCK)
+            self._fifo_stderr = os.open(self._fifo_stderr_name, os.O_RDONLY|os.O_NONBLOCK)            
 
     def _send(self, s, mesg):
         s.send(mesg + chr(0))
@@ -52,12 +54,16 @@ class Python(object):
         buf = ''
         while 1:
             try:
-                buf += os.read(self._fifo, self._bufsize)
+                s = os.read(self._fifo, self._bufsize)
+                buf += s
             except OSError:
                 # nothing more available to read
                 return buf
+            #if not s:
+            #    return buf
 
-    def _read_next_mesg_noblock(self):
+
+    def _recv_next_mesg_noblock(self):
         self._stream += self.recv()
         i = self._stream.find(chr(0))
         if i == -1:
@@ -66,13 +72,15 @@ class Python(object):
         self._stream = self._stream[i+1:]
         return mesg
 
-    def next_mesg(self, block=True):
+    def recv_mesg(self, block=True):
         while 1:
-            m = self._read_next_mesg_noblock()
+            m = self._recv_next_mesg_noblock()
             if not block or m is not None:
                 return m
 
     def __del__(self):
+        if not hasattr(self, '_child_pid'):
+            return
         try:
             os.unlink(self._fifo_name)
         except:
